Convert BlockBuilder into createBlocks
