+++ src/components/Text/Annotated/core/block/index.ts
@@ -1,3 +1,3 @@
-export { BlockBuilder } from "./BlockBuilder.ts";
+export { createBlocks } from "./createBlocks.ts";
 export type { Element, Block, Inline } from "./BlockModel.ts";
 export type { BlockSchema } from "./BlockSchema.ts";
