+++ src/components/Text/Annotated/core/index.ts
@@ -31,6 +31,6 @@ export {
 export { TextSegmentsViewer } from "./inline/TextSegmentsViewer.tsx";
 
 export type { BlockType, BlockAnnotationSegment } from "./AnnotationModel.ts";
-export { BlockBuilder, type BlockSchema } from "./block";
+export { createBlocks, type BlockSchema } from "./block";
 export type { Element, Block, Inline } from "./block";
 export { assignGroupToSegments } from "./utils/assignGroupToSegments.ts";
